Forgot to commit earlier. Almost finished but tableView isn't populating

+++ ObjC-Document/ObjC-Document/NSString+KTMWordCount.h
@@ -12,7 +12,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface NSString (KTMWordCount)
 
-- (int)ktm_wordCount;
+- (NSUInteger)ktm_wordCount;
 
 @end
